initial logic for parsing value
I'm imagining a special "value" (in this case ::) that would designated that a value has been decorated. Then the left handside of the decoration would contain the instructions to be applied (separated by another special value, maybe & by default?).. this would allow specifying information (like type or path) for a given value to be applied. I'd like to eventually be able to have the user specify decorator keys outside the main file, that could also be applied when parsing